defconfig: disable fortification
Since gcc-12 master has further tightened inlining restrictions, the builtin functions are all feaking out about potential overflows.
Some are being triggered as false positives and would have unintended behavior if it's attempted to be fixed.
It's better to leave them as is instead of attempting to fix these as it can further break things unnecessarily.

Signed-off-by: Reinazhard <reinazhard@gmail.com>